Renamed Range to ILRange to avoid conflicting with the Net Core type.
